[CR] Shop model code reviewed
