+++ web/etsy/etsy_core/models/__init__.py
@@ -1,1 +1,2 @@
 from .user import User
+from .shop import Shop
